Handle latex errors when converting to dvi
Whenever latex fails (for example due to a missing package) the code
continue and ends up failing due to a missing svg file.

Handle latex non 0 exit code.
Pass to latex -halt-on-error so it does not attempt to ask user for a
file, that is ignored in batch mode anyway.

Write the latex output to stderr and raise an exception to abort early.

Signed-off-by: Antoine Musso <hashar@free.fr>