manimlib/mobject/mobject: use deepcopy by default
Using a shallow copy can result in duplicate objects for references in
an mobject that aren't part of the submobjects array, see [1].
Deepcopy fixes this and doesn't seem to have any other adverse
drawbacks.

Use it by default, however add a config value allowing this to be
overwritten for backwards compatibility.

[1]: https://github.com/3b1b/manim/issues/688